requirements.txt
Django
numpy
djangorestframework